requirements.txt
python-telegram-bot
logging
tweepy